# QL_FOODDELI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QL_FOODDELI API (Dapper & SignalR)",
    "description": "Hệ thống backend API quản lý giao đồ ăn nhanh QL_FOODDELI - Sprint 5",
    "version": "v1"
  },
  "paths": {
...
    "/api/DonHang": {
      "get": {
        "tags": [
          "DonHang"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "DonHang"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DonHangCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DonHangCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DonHangCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "DonHang"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DonHangUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DonHangUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DonHangUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DonHang/{id}": {
      "get": {
        "tags": [
          "DonHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "DonHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/DonHang/{id}/lich-su-trang-thai": {
      "get": {
        "tags": [
          "DonHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/DonHang/cap-nhat-trang-thai/{id}": {
      "put": {
        "tags": [
          "DonHang"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trangThaiMoi",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ghiChu",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChiTietDonHangCreateDto": {
        "type": "object",
        "properties": {
          "maChiTietDonHang": {
            "type": "string",
            "nullable": true
          },
          "maMonAn": {
            "type": "string",
            "nullable": true
          },
          "soLuong": {
            "type": "integer",
            "format": "int32"
          },
          "donGia": {
            "type": "number",
            "format": "double"
          },
          "thanhTien": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ChiTietDonHangUpdateDto": {
        "type": "object",
        "properties": {
          "maChiTietDonHang": {
            "type": "string",
            "nullable": true
          },
          "maMonAn": {
            "type": "string",
            "nullable": true
          },
          "soLuong": {
            "type": "integer",
            "format": "int32"
          },
          "donGia": {
            "type": "number",
            "format": "double"
          },
          "thanhTien": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DonHangCreateRequest": {
        "type": "object",
        "properties": {
          "maDonHang": {
            "type": "string",
            "nullable": true
          },
          "maDiaChi": {
            "type": "string",
            "nullable": true
          },
          "tongTien": {
            "type": "number",
            "format": "double"
          },
          "phiGiaoHang": {
            "type": "number",
            "format": "double"
          },
          "giamGia": {
            "type": "number",
            "format": "double"
          },
          "thanhTien": {
            "type": "number",
            "format": "double"
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          },
          "chiTiets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChiTietDonHangCreateDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DonHangUpdateRequest": {
        "type": "object",
        "properties": {
          "maDonHang": {
            "type": "string",
            "nullable": true
          },
          "trangThai": {
            "type": "string",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          },
          "chiTiets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChiTietDonHangUpdateDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Dán mã JWT Token thu được sau khi Đăng nhập vào ô này (KHÔNG gõ chữ Bearer):",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}